+++ docs/changelog.md
@@ -17,6 +17,7 @@ category: Administration
 * feat: ensure compatibility with Jira 10.4.0
 * refactor: use Java 16 syntax for lists
 * refactor: extract logic for custom field name syntax
+* test: improve tests for custom field name logic
 
 ### [24.11.0] - 2024-11-11